GRAPHICS: Moved SIMD code to new translation unit
