+++ graphics/blit/blit-alpha.cpp
@@ -21,11 +21,6 @@
 
 #include "graphics/blit.h"
 #include "graphics/pixelformat.h"
-#include "common/system.h"
-
-#include "graphics/blit/blit-neon.cpp"
-#include "graphics/blit/blit-sse2.cpp"
-#include "graphics/blit/blit-avx2.cpp"
 
 namespace Graphics {
 
@@ -171,512 +166,5 @@ bool setAlpha(byte *dst, const byte *src,
 
     return true;
 }
-
-
-BlendBlit::Args::Args(byte *dst, const byte *src,
-    const uint _dstPitch, const uint _srcPitch,
-    const int posX, const int posY,
-    const uint _width, const uint _height,
-    const int _scaleX, const int _scaleY,
-    const int scaleXsrcOff, const int scaleYsrcOff,
-    const uint32 colorMod, const uint _flipping) :
-        xp(0), yp(0), dstPitch(_dstPitch),
-        width(_width), height(_height), color(colorMod),
-        scaleX(_scaleX), scaleY(_scaleY), flipping(_flipping),
-        scaleXoff(scaleXsrcOff), scaleYoff(scaleYsrcOff) {
-    bool doScale = scaleX != SCALE_THRESHOLD || scaleY != SCALE_THRESHOLD;
-    
-    rgbmod   = ((colorMod & kRGBModMask) != kRGBModMask);
-    alphamod = ((colorMod & kAModMask)   != kAModMask);
-    inStep = 4;
-    inoStep = _srcPitch;
-    if (flipping & FLIP_H) {
-        inStep = -inStep;
-        xp = width - 1;
-        if (doScale) xp = xp * scaleX / SCALE_THRESHOLD;
-    }
-
-    if (flipping & FLIP_V) {
-        inoStep = -inoStep;
-        yp = height - 1;
-        if (doScale) yp = yp * scaleY / SCALE_THRESHOLD;
-    }
-
-    ino = src + yp * _srcPitch + xp * 4;
-    outo = dst + posY * _dstPitch + posX * 4;
-}
-
-/**
- * Optimized version of doBlit to be used with multiply blended blitting
- */
-template<bool doscale, bool rgbmod, bool alphamod>
-void BlendBlit::doBlitMultiplyBlendLogicGeneric(Args &args) {
-    const byte *in;
-    byte *out;
-
-    int scaleXCtr, scaleYCtr = args.scaleYoff;
-    const byte *inBase;
-
-    const byte rawcr = (args.color >> kRModShift) & 0xFF;
-    const byte rawcg = (args.color >> kGModShift) & 0xFF;
-    const byte rawcb = (args.color >> kBModShift) & 0xFF;
-    const byte ca = alphamod ? ((args.color >> kAModShift) & 0xFF) : 255;
-    const uint32 cr = rgbmod   ? (rawcr == 255 ? 256 : rawcr) : 256;
-    const uint32 cg = rgbmod   ? (rawcg == 255 ? 256 : rawcg) : 256;
-    const uint32 cb = rgbmod   ? (rawcb == 255 ? 256 : rawcb) : 256;
-
-    for (uint32 i = 0; i < args.height; i++) {
-        if (doscale) {
-            inBase = args.ino + scaleYCtr / SCALE_THRESHOLD * args.inoStep;
-            scaleXCtr = args.scaleXoff;
-        } else {
-            in = args.ino;
-        }
-        out = args.outo;
-        for (uint32 j = 0; j < args.width; j++) {
-            if (doscale) {
-                in = inBase + scaleXCtr / SCALE_THRESHOLD * args.inStep;
-            }
-
-            uint32 ina = in[kAIndex] * ca >> 8;
-
-            if (ina != 0) {
-                out[kBIndex] = out[kBIndex] * ((in[kBIndex] * cb * ina) >> 16) >> 8;
-                out[kGIndex] = out[kGIndex] * ((in[kGIndex] * cg * ina) >> 16) >> 8;
-                out[kRIndex] = out[kRIndex] * ((in[kRIndex] * cr * ina) >> 16) >> 8;
-            }
-
-            if (doscale)
-                scaleXCtr += args.scaleX;
-            else
-                in += args.inStep;
-            out += 4;
-        }
-        if (doscale)
-            scaleYCtr += args.scaleY;
-        else
-            args.ino += args.inoStep;
-        args.outo += args.dstPitch;
-    }
-
-}
-
-template<bool doscale, bool rgbmod, bool alphamod>
-void BlendBlit::doBlitAlphaBlendLogicGeneric(Args &args) {
-    const byte *in;
-    byte *out;
-
-    int scaleXCtr, scaleYCtr = args.scaleYoff;
-    const byte *inBase;
-
-    const byte ca = alphamod ? ((args.color >> kAModShift) & 0xFF) : 255;
-    const byte cr = rgbmod   ? ((args.color >> kRModShift) & 0xFF) : 255;
-    const byte cg = rgbmod   ? ((args.color >> kGModShift) & 0xFF) : 255;
-    const byte cb = rgbmod   ? ((args.color >> kBModShift) & 0xFF) : 255;
-
-    for (uint32 i = 0; i < args.height; i++) {
-        if (doscale) {
-            inBase = args.ino + scaleYCtr / SCALE_THRESHOLD * args.inoStep;
-            scaleXCtr = args.scaleXoff;
-        } else {
-            in = args.ino;
-        }
-        out = args.outo;
-        for (uint32 j = 0; j < args.width; j++) {
-            if (doscale) {
-                in = inBase + scaleXCtr / SCALE_THRESHOLD * args.inStep;
-            }
-
-            uint32 ina = in[kAIndex] * ca >> 8;
-
-            if (ina != 0) {
-                if (rgbmod) {
-                    const uint outb = (out[kBIndex] * (255 - ina) >> 8);
-                    const uint outg = (out[kGIndex] * (255 - ina) >> 8);
-                    const uint outr = (out[kRIndex] * (255 - ina) >> 8);
-
-                    out[kAIndex] = 255;
-                    out[kBIndex] = outb + (in[kBIndex] * ina * cb >> 16);
-                    out[kGIndex] = outg + (in[kGIndex] * ina * cg >> 16);
-                    out[kRIndex] = outr + (in[kRIndex] * ina * cr >> 16);
-                } else {
-                    out[kAIndex] = 255;
-                    out[kBIndex] = (out[kBIndex] * (255 - ina) + in[kBIndex] * ina) >> 8;
-                    out[kGIndex] = (out[kGIndex] * (255 - ina) + in[kGIndex] * ina) >> 8;
-                    out[kRIndex] = (out[kRIndex] * (255 - ina) + in[kRIndex] * ina) >> 8;
-                    
-                }
-            }
-
-            if (doscale)
-                scaleXCtr += args.scaleX;
-            else
-                in += args.inStep;
-            out += 4;
-        }
-
-        if (doscale)
-            scaleYCtr += args.scaleY;
-        else
-            args.ino += args.inoStep;
-        args.outo += args.dstPitch;
-    }
-}
-
-/**
- * Optimized version of doBlit to be used with subtractive blended blitting
- */
-template<bool doscale, bool rgbmod>
-void BlendBlit::doBlitSubtractiveBlendLogicGeneric(Args &args) {
-    const byte *in;
-    byte *out;
-
-    int scaleXCtr, scaleYCtr = args.scaleYoff;
-    const byte *inBase;
-
-    const byte rawcr = (args.color >> kRModShift) & 0xFF;
-    const byte rawcg = (args.color >> kGModShift) & 0xFF;
-    const byte rawcb = (args.color >> kBModShift) & 0xFF;
-    const uint32 cr = rgbmod   ? (rawcr == 255 ? 256 : rawcr) : 256;
-    const uint32 cg = rgbmod   ? (rawcg == 255 ? 256 : rawcg) : 256;
-    const uint32 cb = rgbmod   ? (rawcb == 255 ? 256 : rawcb) : 256;
-
-    for (uint32 i = 0; i < args.height; i++) {
-        if (doscale) {
-            inBase = args.ino + scaleYCtr / SCALE_THRESHOLD * args.inoStep;
-            scaleXCtr = args.scaleXoff;
-        } else {
-            in = args.ino;
-        }
-        out = args.outo;
-        for (uint32 j = 0; j < args.width; j++) {
-            if (doscale) {
-                in = inBase + scaleXCtr / SCALE_THRESHOLD * args.inStep;
-            }
-
-            out[kAIndex] = 255;
-            out[kBIndex] = MAX<int32>(out[kBIndex] - ((in[kBIndex] * cb  * (out[kBIndex]) * in[kAIndex]) >> 24), 0);
-            out[kGIndex] = MAX<int32>(out[kGIndex] - ((in[kGIndex] * cg  * (out[kGIndex]) * in[kAIndex]) >> 24), 0);
-            out[kRIndex] = MAX<int32>(out[kRIndex] - ((in[kRIndex] * cr * (out[kRIndex]) * in[kAIndex]) >> 24), 0);
-
-            if (doscale)
-                scaleXCtr += args.scaleX;
-            else
-                in += args.inStep;
-            out += 4;
-        }
-        if (doscale)
-            scaleYCtr += args.scaleY;
-        else
-            args.ino += args.inoStep;
-        args.outo += args.dstPitch;
-    }
-}
-
-/**
- * Optimized version of doBlit to be used with additive blended blitting
- */
-template<bool doscale, bool rgbmod, bool alphamod>
-void BlendBlit::doBlitAdditiveBlendLogicGeneric(Args &args) {
-    const byte *in;
-    byte *out;
-
-    int scaleXCtr, scaleYCtr = args.scaleYoff;
-    const byte *inBase;
-
-    const byte rawcr = (args.color >> kRModShift) & 0xFF;
-    const byte rawcg = (args.color >> kGModShift) & 0xFF;
-    const byte rawcb = (args.color >> kBModShift) & 0xFF;
-    const byte ca = alphamod ? ((args.color >> kAModShift) & 0xFF) : 255;
-    const uint32 cr = rgbmod   ? (rawcr == 255 ? 256 : rawcr) : 256;
-    const uint32 cg = rgbmod   ? (rawcg == 255 ? 256 : rawcg) : 256;
-    const uint32 cb = rgbmod   ? (rawcb == 255 ? 256 : rawcb) : 256;
-
-    for (uint32 i = 0; i < args.height; i++) {
-        if (doscale) {
-            inBase = args.ino + scaleYCtr / SCALE_THRESHOLD * args.inoStep;
-            scaleXCtr = args.scaleXoff;
-        } else {
-            in = args.ino;
-        }
-        out = args.outo;
-        for (uint32 j = 0; j < args.width; j++) {
-            if (doscale) {
-                in = inBase + scaleXCtr / SCALE_THRESHOLD * args.inStep;
-            }
-
-            uint32 ina = in[kAIndex] * ca >> 8;
-
-            if (ina != 0) {
-                out[kBIndex] = out[kBIndex] + ((in[kBIndex] * cb * ina) >> 16);
-                out[kGIndex] = out[kGIndex] + ((in[kGIndex] * cg * ina) >> 16);
-                out[kRIndex] = out[kRIndex] + ((in[kRIndex] * cr * ina) >> 16);
-            }
-
-            if (doscale)
-                scaleXCtr += args.scaleX;
-            else
-                in += args.inStep;
-            out += 4;
-        }
-
-        if (doscale)
-            scaleYCtr += args.scaleY;
-        else
-            args.ino += args.inoStep;
-        args.outo += args.dstPitch;
-    }
-}
-
-template<bool doscale>
-void BlendBlit::doBlitOpaqueBlendLogicGeneric(Args &args) {
-    const byte *in;
-    byte *out;
-
-    int scaleXCtr, scaleYCtr = args.scaleYoff;
-    const byte *inBase;
-
-    for (uint32 i = 0; i < args.height; i++) {
-        if (doscale) {
-            inBase = args.ino + (scaleYCtr + 1) / SCALE_THRESHOLD * args.inoStep;
-            scaleXCtr = args.scaleXoff;
-        } else {
-            in = args.ino;
-        }
-        out = args.outo;
-
-        if (doscale) {
-            for (uint32 j = 0; j < args.width; j++) {
-                in = inBase + scaleXCtr / SCALE_THRESHOLD * args.inStep;
-                *(uint32 *)out = *(const uint32 *)in | kAModMask;
-                scaleXCtr += args.scaleX;
-                out += 4;
-            }
-        } else {
-            for (uint32 j = 0; j < args.width; j++) {
-                *(uint32 *)out = *(const uint32 *)in | kAModMask;
-                in += args.inStep;
-                out += 4;
-            }
-        }
-
-        if (doscale)
-            scaleYCtr += args.scaleY;
-        else
-            args.ino += args.inoStep;
-        args.outo += args.dstPitch;
-    }
-}
-
-template<bool doscale>
-void BlendBlit::doBlitBinaryBlendLogicGeneric(Args &args) {
-    const byte *in;
-    byte *out;
-
-    int scaleXCtr, scaleYCtr = args.scaleYoff;
-    const byte *inBase;
-
-    for (uint32 i = 0; i < args.height; i++) {
-        if (doscale) {
-            inBase = args.ino + scaleYCtr / SCALE_THRESHOLD * args.inoStep;
-            scaleXCtr = args.scaleXoff;
-        } else {
-            in = args.ino;
-        }
-        out = args.outo;
-        for (uint32 j = 0; j < args.width; j++) {
-            if (doscale) {
-                in = inBase + scaleXCtr / SCALE_THRESHOLD * args.inStep;
-            }
-
-            uint32 pix = *(const uint32 *)in, pixout = *(const uint32 *)out;
-            uint32 mask = (pix & kAModMask) ? 0xffffffff : 0;
-            pixout &= ~mask;
-            pix = (pix | kAModMask) & mask;
-            *(uint32 *)out = pixout | pix;
-            
-            if (doscale)
-                scaleXCtr += args.scaleX;
-            else
-                in += args.inStep;
-            out += 4;
-        }
-        if (doscale)
-            scaleYCtr += args.scaleY;
-        else
-            args.ino += args.inoStep;
-        args.outo += args.dstPitch;
-    }
-}
-
-// Initialize this to nullptr at the start
-BlendBlit::BlitFunc BlendBlit::blitFunc = nullptr;
-
-// Only blits to and from 32bpp images
-// So this function is just here to jump to whatever function is in
-// BlendBlit::blitFunc. This way, we can detect at runtime whether or not
-// the cpu has certain SIMD feature enabled or not.
-void BlendBlit::blit(byte *dst, const byte *src,
-                     const uint dstPitch, const uint srcPitch,
-                     const int posX, const int posY,
-                     const uint width, const uint height,
-                     const int scaleX, const int scaleY,
-                     const int scaleXsrcOff, const int scaleYsrcOff,
-                     const uint32 colorMod, const uint flipping,
-                     const TSpriteBlendMode blendMode,
-                     const AlphaType alphaType) {
-    if (width == 0 || height == 0) return;
-
-    // If no function has been selected yet, detect and select
-    if (!blitFunc) {
-        // Get the correct blit function
-        blitFunc = blitGeneric;
-#ifdef SCUMMVM_NEON
-        if (g_system->hasFeature(OSystem::kFeatureCpuNEON)) blitFunc = blitNEON;
-#endif
-#ifdef SCUMMVM_SSE2
-        if (g_system->hasFeature(OSystem::kFeatureCpuSSE2)) blitFunc = blitSSE2;
-#endif
-#ifdef SCUMMVM_AVX2
-        if (g_system->hasFeature(OSystem::kFeatureCpuAVX2)) blitFunc = blitAVX2;
-#endif
-    }
-    
-    Args args(dst, src, dstPitch, srcPitch, posX, posY, width, height, scaleX, scaleY, scaleXsrcOff, scaleYsrcOff, colorMod, flipping);
-    blitFunc(args, blendMode, alphaType);
-}
-
-// This is just a macro to expand it because its a pretty simple function where
-// readabiliy doesn't matter too much and macros tend to work faster better than functors
-#define BLIT_FUNC(ext) \
-    void BlendBlit::blit##ext(Args &args, const TSpriteBlendMode &blendMode, const AlphaType &alphaType) { \
-        bool rgbmod   = ((args.color & kRGBModMask) != kRGBModMask); \
-        bool alphamod = ((args.color & kAModMask)   != kAModMask); \
-        if (args.scaleX == SCALE_THRESHOLD && args.scaleY == SCALE_THRESHOLD) { \
-            if (args.color == 0xffffffff && blendMode == BLEND_NORMAL && alphaType == ALPHA_OPAQUE) { \
-                doBlitOpaqueBlendLogic##ext<false>(args); \
-            } else if (args.color == 0xffffffff && blendMode == BLEND_NORMAL && alphaType == ALPHA_BINARY) { \
-                doBlitBinaryBlendLogic##ext<false>(args); \
-            } else { \
-                if (blendMode == BLEND_ADDITIVE) { \
-                    if (rgbmod) { \
-                        if (alphamod) { \
-                            doBlitAdditiveBlendLogic##ext<false, true, true>(args); \
-                        } else { \
-                            doBlitAdditiveBlendLogic##ext<false, true, false>(args); \
-                        } \
-                    } else { \
-                        if (alphamod) { \
-                            doBlitAdditiveBlendLogic##ext<false, false, true>(args); \
-                        } else { \
-                            doBlitAdditiveBlendLogic##ext<false, false, false>(args); \
-                        } \
-                    } \
-                } else if (blendMode == BLEND_SUBTRACTIVE) { \
-                    if (rgbmod) { \
-                        doBlitSubtractiveBlendLogic##ext<false, true>(args); \
-                    } else { \
-                        doBlitSubtractiveBlendLogic##ext<false, false>(args); \
-                    } \
-                } else if (blendMode == BLEND_MULTIPLY) { \
-                    if (rgbmod) { \
-                        if (alphamod) { \
-                            doBlitMultiplyBlendLogic##ext<false, true, true>(args); \
-                        } else { \
-                            doBlitMultiplyBlendLogic##ext<false, true, false>(args); \
-                        } \
-                    } else { \
-                        if (alphamod) { \
-                            doBlitMultiplyBlendLogic##ext<false, false, true>(args); \
-                        } else { \
-                            doBlitMultiplyBlendLogic##ext<false, false, false>(args); \
-                        } \
-                    } \
-                } else { \
-                    assert(blendMode == BLEND_NORMAL); \
-                    if (rgbmod) { \
-                        if (alphamod) { \
-                            doBlitAlphaBlendLogic##ext<false, true, true>(args); \
-                        } else { \
-                            doBlitAlphaBlendLogic##ext<false, true, false>(args); \
-                        } \
-                    } else { \
-                        if (alphamod) { \
-                            doBlitAlphaBlendLogic##ext<false, false, true>(args); \
-                        } else { \
-                            doBlitAlphaBlendLogic##ext<false, false, false>(args); \
-                        } \
-                    } \
-                } \
-            } \
-        } else { \
-            if (args.color == 0xffffffff && blendMode == BLEND_NORMAL && alphaType == ALPHA_OPAQUE) { \
-                doBlitOpaqueBlendLogic##ext<true>(args); \
-            } else if (args.color == 0xffffffff && blendMode == BLEND_NORMAL && alphaType == ALPHA_BINARY) { \
-                doBlitBinaryBlendLogic##ext<true>(args); \
-            } else { \
-                if (blendMode == BLEND_ADDITIVE) { \
-                    if (rgbmod) { \
-                        if (alphamod) { \
-                            doBlitAdditiveBlendLogic##ext<true, true, true>(args); \
-                        } else { \
-                            doBlitAdditiveBlendLogic##ext<true, true, false>(args); \
-                        } \
-                    } else { \
-                        if (alphamod) { \
-                            doBlitAdditiveBlendLogic##ext<true, false, true>(args); \
-                        } else { \
-                            doBlitAdditiveBlendLogic##ext<true, false, false>(args); \
-                        } \
-                    } \
-                } else if (blendMode == BLEND_SUBTRACTIVE) { \
-                    if (rgbmod) { \
-                        doBlitSubtractiveBlendLogic##ext<true, true>(args); \
-                    } else { \
-                        doBlitSubtractiveBlendLogic##ext<true, false>(args); \
-                    } \
-                } else if (blendMode == BLEND_MULTIPLY) { \
-                    if (rgbmod) { \
-                        if (alphamod) { \
-                            doBlitMultiplyBlendLogic##ext<true, true, true>(args); \
-                        } else { \
-                            doBlitMultiplyBlendLogic##ext<true, true, false>(args); \
-                        } \
-                    } else { \
-                        if (alphamod) { \
-                            doBlitMultiplyBlendLogic##ext<true, false, true>(args); \
-                        } else { \
-                            doBlitMultiplyBlendLogic##ext<true, false, false>(args); \
-                        } \
-                    } \
-                } else { \
-                    assert(blendMode == BLEND_NORMAL); \
-                    if (rgbmod) { \
-                        if (alphamod) { \
-                            doBlitAlphaBlendLogic##ext<true, true, true>(args); \
-                        } else { \
-                            doBlitAlphaBlendLogic##ext<true, true, false>(args); \
-                        } \
-                    } else { \
-                        if (alphamod) { \
-                            doBlitAlphaBlendLogic##ext<true, false, true>(args); \
-                        } else { \
-                            doBlitAlphaBlendLogic##ext<true, false, false>(args); \
-                        } \
-                    } \
-                } \
-            } \
-        } \
-    }
-BLIT_FUNC(Generic)
-#ifdef SCUMMVM_NEON
-BLIT_FUNC(NEON)
-#endif
-#ifdef SCUMMVM_SSE2
-BLIT_FUNC(SSE2)
-#endif
-#ifdef SCUMMVM_AVX2
-BLIT_FUNC(AVX2)
-#endif
 
 } // End of namespace Graphics
